requirements.txt
allure-python-commons
allure-pytest
playwright
pytest
pytest-playwright
pytest-rerunfailures
Faker
python-dotenv